- Added support for playerSaveData in custom json data hb_datastore because serialization was being unreliable - changed ManagedPlayer implemenation to save and load data from datastore - added util methods in HBUtil::networkUtil to support translation between json and CompoundTag

+++ common/src/main/java/com/holybuckets/foundation/core/EssenceCauldronManager.java
@@ -147,7 +147,7 @@ public class EssenceCauldronManager {
 
         BlockPos playerEnd = player.blockPosition();
         int blockDist = HBUtil.BlockUtil.distanceSqr(playerStart, playerEnd);
-        String dist = Math.sqrt(blockDist) + "";
+        String dist = (int) Math.sqrt(blockDist) + "";
         CommonClass.MESSAGER.sendChat(player,
             Component.translatable("item.hbs_foundation.enchanted_essence.teleport_success_chat",
                 HBUtil.BlockUtil.positionToString(playerStart),
